+++ libdirections-offboard/build.gradle
@@ -19,8 +19,10 @@ dependencies {
 
     api(project(':libnavigation-base'))
 
+    implementation dependenciesList.kotlinStdLib
+
     //ktlint
     ktlint dependenciesList.ktlint
 
-    implementation dependenciesList.kotlinStdLib
+    implementation dependenciesList.mapboxSdkServices
 }
